requirements.txt
aiohttp
anytree
appdirs
aria2p
attrdict
beautifulsoup4
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
heroku3
js2py
lxml
megasdkrestclient
psutil
psycopg2-binary
pybase64
Pyrogram
python-dotenv
python-magic
python-telegram-bot
pytz
quoters
requests
speedtest-cli
telegraph
tenacity
TgCrypto
urllib3
yt-dlp

